Add new `ConsumerSettings::disableOptimisticResume` setting to allow users to disable the "optimistic resume" optimisation && Write benchmarks where this optimisation is disabled
