+++ zio-kafka/src/main/scala/zio/kafka/consumer/internal/Runloop.scala
@@ -5,10 +5,10 @@ import org.apache.kafka.common.TopicPartition
 import org.apache.kafka.common.errors.RebalanceInProgressException
 import zio._
 import zio.kafka.consumer.Consumer.{ OffsetRetrieval, RunloopTimeout }
+import zio.kafka.consumer._
 import zio.kafka.consumer.diagnostics.{ DiagnosticEvent, Diagnostics }
 import zio.kafka.consumer.internal.ConsumerAccess.ByteArrayKafkaConsumer
 import zio.kafka.consumer.internal.Runloop._
-import zio.kafka.consumer.{ CommittableRecord, RebalanceConsumer, RebalanceListener, Subscription }
 import zio.stream._
 
 import java.util
@@ -27,7 +27,8 @@ private[consumer] final class Runloop private (
   offsetRetrieval: OffsetRetrieval,
   userRebalanceListener: RebalanceListener,
   restartStreamsOnRebalancing: Boolean,
-  currentStateRef: Ref[State]
+  currentStateRef: Ref[State],
+  consumerSettings: ConsumerSettings
 ) {
 
   private def newPartitionStream(tp: TopicPartition): UIO[PartitionStreamControl] =
@@ -241,11 +242,15 @@ private[consumer] final class Runloop private (
     requestedPartitions: Set[TopicPartition],
     streams: Chunk[PartitionStreamControl]
   ): Unit = {
-    val resumeTps = new java.util.ArrayList[TopicPartition](streams.size)
-    val pauseTps  = new java.util.ArrayList[TopicPartition](streams.size)
-    streams.foreach { stream =>
-      val tp       = stream.tp
-      val toResume = requestedPartitions.contains(tp) || stream.optimisticResume
+    val resumeTps               = new java.util.ArrayList[TopicPartition](streams.size)
+    val pauseTps                = new java.util.ArrayList[TopicPartition](streams.size)
+    val optimisticResumeEnabled = !consumerSettings.disableOptimisticResume
+    val iterator                = streams.iterator
+    while (iterator.hasNext) {
+      val stream = iterator.next()
+      val tp     = stream.tp
+      val toResume =
+        requestedPartitions.contains(tp) || (optimisticResumeEnabled && stream.optimisticResume)
       if (toResume) resumeTps.add(tp) else pauseTps.add(tp)
       stream.addPollHistory(toResume)
     }
@@ -537,7 +542,8 @@ private[consumer] object Runloop {
     offsetRetrieval: OffsetRetrieval,
     userRebalanceListener: RebalanceListener,
     restartStreamsOnRebalancing: Boolean,
-    runloopTimeout: Duration
+    runloopTimeout: Duration,
+    consumerSettings: ConsumerSettings
   ): ZIO[Scope, Throwable, Runloop] =
     for {
       commandQueue       <- ZIO.acquireRelease(Queue.bounded[RunloopCommand](commandQueueSize))(_.shutdown)
@@ -563,7 +569,8 @@ private[consumer] object Runloop {
                   offsetRetrieval = offsetRetrieval,
                   userRebalanceListener = userRebalanceListener,
                   restartStreamsOnRebalancing = restartStreamsOnRebalancing,
-                  currentStateRef = currentStateRef
+                  currentStateRef = currentStateRef,
+                  consumerSettings = consumerSettings
                 )
       _ <- ZIO.logDebug("Starting Runloop")
